pyproject.toml
[project]
name = "data-quality"
version = "0.1.0"
description = ""
authors = [
    {name = "Juan Veronez",email = "juan.monteirov@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "mkdocs (>=1.6.1,<2.0.0)",
    "mkdocs-mermaid2-plugin (>=1.2.1,<2.0.0)",
    "mkdocs-material (>=9.6.4,<10.0.0)",
    "mkdocstrings[python] (>=0.28.0,<0.29.0)",
    "taskipy (>=1.14.1,<2.0.0)",
    "isort (>=6.0.0,<7.0.0)",
    "black (>=25.1.0,<26.0.0)",
    "pytest (>=8.3.4,<9.0.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "pandera[io] (>=0.22.1,<0.23.0)",
    "sqlalchemy (>=2.0.38,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "duckdb (>=1.2.0,<2.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.taskipy.tasks]
test = "python -m pytest -v"
run = "python app/main.py"
doc = "mkdocs serve"
format = """
isort .
black .
"""